+++ utils/image_convert.py
@@ -7,6 +7,7 @@
 # https://github.com/microsoft/DirectXTex/releases/tag/jan2021
 import math
 import os.path as op
+import shutil
 import subprocess
 
 FORMATS = {
@@ -19,7 +20,7 @@ TEXCONV_PATH = op.join(op.dirname(__file__), 'DirectXTex', 'texconv.exe')
 CONVERT_COMAMND = [TEXCONV_PATH, '-y', '-m', '13', '-f']
 
 
-def convert_image(fpath, type_, size):
+def convert_image(fpath: str, out_fpath: str, type_: str, size: tuple):
     """ Take an image and convert it to .DDS. """
     # We can only generate as many mip maps as the image will allow based on
     # size.
@@ -27,13 +28,16 @@ def convert_image(fpath, type_, size):
         # There is an issue with reading the dimensions... raise an error
         raise ValueError(f'Cannot determine dimensions of {fpath}. '
                          'Please check the image is valid in blender.')
-    out_fpath = op.splitext(fpath)[0] + '.DDS'
     mips = math.floor(math.log(min(size), 2))
     CONVERT_COMAMND[2] = str(mips)
     command = CONVERT_COMAMND + [FORMATS[type_], fpath, '-o',
                                  op.dirname(out_fpath)]
     subprocess.call(command)
-    if op.exists(out_fpath):
-        return out_fpath
-    else:
-        raise FileNotFoundError
+    if not op.exists(out_fpath):
+        # In this case the actual filename may be different. Move the converted
+        # file to have this new file name.
+        shutil.move(
+            op.join(op.dirname(out_fpath),
+                    op.splitext(op.basename(fpath))[0]) + '.DDS',
+            out_fpath)
+    return out_fpath